+++ configs/sac.py
@@ -1,4 +1,4 @@
-from oprl.algos import OffPolicyAlgorithm
+from oprl.algos.protocols import OffPolicyAlgorithm
 from oprl.algos.sac import SAC
 from oprl.parse_args import parse_args
 from oprl.logging import (
@@ -8,6 +8,7 @@ from oprl.logging import (
 )
 set_logging()
 from oprl.environment import make_env as _make_env
+from oprl.buffers.episodic_buffer import ReplayBufferProtocol, EpisodicReplayBuffer
 from oprl.train import run_training
 
 args = parse_args()
@@ -54,6 +55,23 @@ make_logger = make_text_logger_func(
 )
 
 
+def make_replay_buffer() -> ReplayBufferProtocol:
+    return EpisodicReplayBuffer(
+        buffer_size=max(config["num_steps"], int(1e6)),
+        state_dim=STATE_DIM,
+        action_dim=ACTION_DIM,
+        device=config["device"],
+    ).create()
+
+
 if __name__ == "__main__":
     args = parse_args()
-    run_training(make_algo, make_env, make_logger, config, args.seeds, args.start_seed)
+    run_training(
+        make_algo=make_algo,
+        make_env=make_env,
+        make_replay_buffer=make_replay_buffer,
+        make_logger=make_logger,
+        config=config,
+        seeds=args.seeds,
+        start_seed=args.start_seed
+    )
