+++ configs/td3.py
@@ -1,6 +1,7 @@
-from oprl.algos import OffPolicyAlgorithm
+from oprl.algos.protocols import OffPolicyAlgorithm
 from oprl.algos.td3 import TD3
 from oprl.parse_args import parse_args
+from oprl.buffers.episodic_buffer import ReplayBufferProtocol, EpisodicReplayBuffer
 from oprl.logging import (
     set_logging,
     LoggerProtocol,
@@ -46,6 +47,15 @@ def make_algo(logger: LoggerProtocol) -> OffPolicyAlgorithm:
     ).create()
 
 
+def make_replay_buffer() -> ReplayBufferProtocol:
+    return EpisodicReplayBuffer(
+        buffer_size=max(config["num_steps"], int(1e6)),
+        state_dim=STATE_DIM,
+        action_dim=ACTION_DIM,
+        device=config["device"],
+    ).create()
+
+
 make_logger = make_text_logger_func(
     config=config,
     algo="TD3",
@@ -55,4 +65,12 @@ make_logger = make_text_logger_func(
 
 if __name__ == "__main__":
     args = parse_args()
-    run_training(make_algo, make_env, make_logger, config, args.seeds, args.start_seed)
+    run_training(
+        make_algo=make_algo,
+        make_env=make_env,
+        make_logger=make_logger,
+        make_replay_buffer=make_replay_buffer,
+        config=config, 
+        seeds=args.seeds, 
+        start_seed=args.start_seed
+    )
