+++ src/oprl/algos/protocols.py
@@ -1,0 +1,24 @@
+from typing import Protocol
+
+import numpy.typing as npt
+
+import torch as t
+
+
+class OffPolicyAlgorithm(Protocol):
+    def create(self) -> "OffPolicyAlgorithm": ...
+
+    def update(
+        self,
+        state: t.Tensor,
+        action: t.Tensor,
+        reward: t.Tensor,
+        done: t.Tensor,
+        next_state: t.Tensor
+    ) -> None: ...
+
+
+class PolicyProtocol(Protocol):
+    def explore(self, state: npt.ArrayLike) -> npt.NDArray: ...
+
+    def exploit(self, state: npt.ArrayLike) -> npt.NDArray: ...
