+++ src/oprl/environment/__init__.py
@@ -1,4 +1,4 @@
-from oprl.environment.protocol import EnvProtocol
+from oprl.environment.protocols import EnvProtocol
 from oprl.environment.dm_control import DMControlEnv
 from oprl.environment.safety_gymnasium import SafetyGym
 from oprl.environment.make_env import make_env
