+++ src/oprl/trainers/protocols.py
@@ -1,0 +1,7 @@
+from typing import Protocol
+
+class TrainerProtocol(Protocol):
+    def train(self) -> None: ...
+
+    def evaluate(self) -> dict[str, float]: ...
+
